html/semantics/forms/the-input-element/type-change-state.html is failing in WebKit https://bugs.webkit.org/show_bug.cgi?id=243311
Reviewed by NOBODY (OOPS!).

Per the specification [1]:
```
If the previous state of the element's type attribute put the value IDL
attribute in the value mode, and the element's value is not the empty string,
and the new state of the element's type attribute puts the value IDL attribute
in either the default mode or the default/on mode, then set the element's value
content attribute to the element's value.
```

In our implementation, we were missing the `and the element's value is not the
empty string` part of the check. As a result, when changing the type from
"number" to "checkbox" (for example) and if m_valueIfDirty was the emptyString,
we would set the "value" content attribute to the empty string. In general, this
is not very observable because input.value would return the empty string if
whether the "value" content attribute is missing or the empty string. However,
for input types such as "checkbox" and "radio", it is observable since
`input.value` is expected to "on" when the "value" content attribute is missing.
This bug would cause `input.value` to return "" instead of "on" in such case.

[1] https://html.spec.whatwg.org/#the-input-element:attr-input-type-37 (step 1)

* LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/type-change-state-expected.txt:
* Source/WebCore/html/HTMLInputElement.cpp:
(WebCore::HTMLInputElement::updateType):