Remove code guarded by USE_CTFONTHASTABLE https://bugs.webkit.org/show_bug.cgi?id=276104 rdar://130928732
Reviewed by NOBODY (OOPS!).

We no longer need this since USE(CTFONTHASTABLE)
is always true for our builds now.

We are also removing the wrapper `fontHasTable` as
it is now exactly what the lower level CT function
`fontHasTable` does. Wrapping seems unnecessary as
the function is just used locally and if any unlikely
change would occur with the CT API it would be fairly
simple to update it.

* Source/WTF/wtf/PlatformUse.h:
* Source/WebCore/platform/graphics/Font.h:
* Source/WebCore/platform/graphics/coretext/FontCoreText.cpp:
(WebCore::fontHasEitherTable):
(WebCore::Font::platformInit):
(WebCore::fontHasTable): Deleted.