(286869@main): <details> disclosure triangle is now tiny with `system-ui` https://bugs.webkit.org/show_bug.cgi?id=285480 rdar://141769356
Reviewed by NOBODY (OOPS!).

On 280433@main when fixing the symbols according to the direction we have we have aligned
the symbol used for each direction. We have chosen to align them with the small variant as
specification [1] let this open and there is no consistency amoung different engines for
different directions.

Here we are choosing to align to the bigger variant of the disclosure triangles as it improves
readability.

* LayoutTests/imported/w3c/web-platform-tests/css/css-counter-styles/counter-style-at-rule/disclosure-styles-expected.html:
* LayoutTests/imported/w3c/web-platform-tests/css/css-counter-styles/counter-style-at-rule/disclosure-styles-ref.html:
* Source/WTF/wtf/unicode/CharacterNames.h:
* Source/WebCore/css/CSSCounterStyle.cpp:
(WebCore::CSSCounterStyle::counterForSystemDisclosureClosed):
(WebCore::CSSCounterStyle::counterForSystemDisclosureOpen):