Avoid requesting CTFontGetBoundingRectsForGlyphs calculate individual boundingRects that are never used https://bugs.webkit.org/show_bug.cgi?id=287587 rdar://144734746
Reviewed by NOBODY (OOPS!).

When calling CTFontGetBoundingRectsForGlyphs we can provide a out ref to boundingRects which is:

"An array of count number of CGRects to receive the computed glyph rects.
Can be NULL, in which case only the overall bounding rect is calculated."

The calls changed on this patch request the calculation of individual bounding rects
but they are never used. Therefore, we can pass NULL here and avoid the computation.

* Source/WebCore/platform/graphics/coretext/FontCoreText.cpp:
(WebCore::Font::platformBoundsForGlyph const):
(WebCore::Font::isProbablyOnlyUsedToRenderIcons const):