SVG text element with text-anchor: middle is misaligned using icon font https://bugs.webkit.org/show_bug.cgi?id=215321 rdar://66869020
Reviewed by NOBODY (OOPS!).

Summing up advance per each fragment will not always
match the whole content as codepoint to glyph relation
is not 1:1. There are cases in which one codepoint is
shaped as multiple glyphs and there are cases in which
multiple codepoints are shaped as a single glyph, for example.

* Source/WebCore/rendering/svg/SVGTextLayoutEngine.cpp:
(WebCore::SVGTextLayoutEngine::computeCurrentFragmentMetrics):
(WebCore::SVGTextLayoutEngine::recordTextFragment):
* Source/WebCore/rendering/svg/SVGTextLayoutEngine.h: