Vorbis decoder can't be instantiated in the content process on Sonoma machines (and earlier) https://bugs.webkit.org/show_bug.cgi?id=300083 rdar://161878608
Reviewed by Per Arne Vollan.

The AudioToolbox's AudioComponentFetchServerRegistrations appears non-functional
on Sonoma and earlier, which is what is relied on to enable the opus and vorbis
decoder. We re-enable the old code to register the audio decoder right
before setting up the sandbox in the WebProcess.

The tests had been marked as failing and so the problem had been missed.
This became more problematic as we are trying to move the MediaPlayerPrivate
back into the WebProcess and audio decoding was non functional.

Re-enabled tests.
* LayoutTests/platform/ios/TestExpectations: Re-enable tests.
* LayoutTests/platform/mac-wk2/TestExpectations: re-enable tests.
* Source/WTF/wtf/PlatformEnableCocoa.h: Add ENABLE_AUDIO_DECODER_REGISTRATION
* Source/WebKit/WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::WebProcess::initializeSandbox): Register Opus and Vorbis decoder just before setting the sandbox.

Canonical link: https://commits.webkit.org/300936@main