byteStreamTee should not be observable from JS rdar://161883428 https://bugs.webkit.org/show_bug.cgi?id=300087
Reviewed by Chris Dumez.

We were previously relying on promise based reads for byteStreamTee.
But this is observable to JS via Object.prototype.then.
To prevent this, we introduce read requests as defined in the specification.
The read request (ReadRequest for default reading and ReadIntoRequest for byob reading) will be used for both the JS read methods and byteStreamTee.

For the JS read methods, the read requests encapsulate a promise and resolve/reject it as defined in the spec.
See https://streams.spec.whatwg.org/#default-reader-read and https://streams.spec.whatwg.org/#byob-reader-read.

For byteStreamTee, the read requests will be used to fuel the teed streams as per https://streams.spec.whatwg.org/#abstract-opdef-readablebytestreamtee.

In byteStreamTee, we need to queue a micro task, which was done by the previous promise-based mechanism.
To do so, we make StreamTeeState a ContextDestructionObserver, which allows us to get to the event loop.
We also make use of ContextDestructionObserver to clean-up StreamTeeState whenver its context is destroyed, which can help for potential ref cycles.

Test: streams/overriding-then.html
Canonical link: https://commits.webkit.org/300941@main