[cocoa] Set WebMUseRemoteAudioVideoRenderer preference to testable https://bugs.webkit.org/show_bug.cgi?id=299838 rdar://161612789
Reviewed by Jer Noble.

Enable preference for testing, covered by existing webm tests.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebCore/platform/graphics/MediaPlayer.cpp:
(WebCore::WTF_REQUIRES_LOCK): Swap the order of registration of MediaPlayerFactory.
MSE and MP4 are more commonly used as webm. It makes more sense to try those first.
* Source/WebCore/platform/graphics/cocoa/MediaPlayerPrivateWebM.mm:
(WebCore::MediaPlayerPrivateWebM::cancelLoad): We incorrectly set the networkState
to idle when the load was cancelled. When running in the GPUP, it didn't make
a difference as the MediaPlayerPrivateRemote stopped accepting new events from
the GPUP once unloaded.
But when running in the content process, it would cause the element to incorrectly
fire the "suspend" event and causing tests to fail.

Canonical link: https://commits.webkit.org/300974@main