Narrow the return type of WebAssemblyBuiltin::jsWrapper https://bugs.webkit.org/show_bug.cgi?id=300177 rdar://161963940
Reviewed by Yusuke Suzuki.

I noticed that WebAssemblyBuiltin::jsWrapper return type is JSObject*. I don't remember
why it wrote it that way, but it should be JSFunction* to be more specific and better
reflect what that method is about.

This patch does not need tests because it only changes a declared type.

Canonical link: https://commits.webkit.org/301005@main