[Build Speed] Make GraphicsLayer.h less expensive rdar://161829723 https://bugs.webkit.org/show_bug.cgi?id=300033
Reviewed by Ryosuke Niwa.

Prior to this patch, GraphicsLayer.h was the 3rd most expensive header used in
WebCore; it is included 268 times in a WebCore Unified build, and on this
machine, each include of that file took the compiler on avegare 1.3s to parse,
for a total of 5.9m CPU minutes of time spent parsing this header.

The primary cause of GraphicsLayer.h being included in a compliation unit was
that it was included by GraphicsLayerContentsDisplayDelegate.h, which only
needed a forward declaration of an enum defined within the GraphicsLayer class
itself. This enum–along with every other enum defined within GraphicsLayer–was
moved into a new GraphicsLayerEnums.h header, and those enums were pulled back
into GraphicsLayer via a using statement. I then looked at every header which
included GraphicsLayer.h and: converted every use of GraphicsLayer enums to use
the new enums defined in GraphicsLayerEnums.h, added a forward declaration for
GraphicsLayer, and removed the GraphicsLayer.h header where possible.

After this patch, GraphicsLayer.h is now the 67th most expensive header used in
WebCore; it is included 37 times, for a total of 49s CPU seconds of time spent
parsing this header, a reduction of 5.1m.

Canonical link: https://commits.webkit.org/301010@main