Add pipeTo support for readable byte streams https://bugs.webkit.org/show_bug.cgi?id=299438# rdar://problem/161244992
Reviewed by Chris Dumez.

We implement the pipeTo algorithm in C++ using a default reader.
This is the same algorithm as for regular streams so it might replace the JS built-in implementation in a follow-up for regular streams.

The implementation is done via PipeToState which stores the necessary state values needed for pipeTo.
It notably keeps a reference to the source and destination to propagate the errors backward and forward.
It also keeps reference to the reader and writer used for pipeTo.

The implementation follows closely the specification so the pipeTo operations are all done in the thread where live the streams.

Covered by rebased tests.

Canonical link: https://commits.webkit.org/301030@main