AX: [iOS] Date fields shouldn't send focused change notification when picker popup is presented https://bugs.webkit.org/show_bug.cgi?id=269269 rdar://122851275
Reviewed by Tyler Wilcock.

VoiceOver focus wasn't being moved to the date picker's popup due to a focused element changed
notification fired on the date element as the picker was expanded. Native pop up elements will
automatically post this notification in the UI process, so this notification conflicted and
created a race condition.

To prevent this, there is a new flag in AXObjectCache that maintains the state of the picker.
If the picker will open, this flag prevents us from sending the conflicting notification. We
clear this state for new focus changes and when the picker is closed (sent via IPC from the
UI process).

Date and DateTimeLocal input types were also added to elements that should defer their focus
(shouldDeferFocusChange). This is because when this race condition occurs, the focus event
happens prior to the DOMActivate event being dispatched, which prevents us from knowing the
state of the picker before handling the notification. By deferring, we allow the DOM event
to occur in time.

Test: accessibility/ios-simulator/date-picker-focus-notification.html

* LayoutTests/accessibility/ios-simulator/date-picker-focus-notification-expected.txt: Added.
* LayoutTests/accessibility/ios-simulator/date-picker-focus-notification.html: Added.
* Source/WebCore/accessibility/AXObjectCache.cpp:
(WebCore::shouldDeferFocusChange):
(WebCore::AXObjectCache::handleFocusedUIElementChanged):
* Source/WebCore/accessibility/AXObjectCache.h:
* Source/WebCore/html/BaseDateAndTimeInputType.cpp:
(WebCore::BaseDateAndTimeInputType::showPicker):
* Source/WebCore/page/Page.cpp:
(WebCore::Page::clearIsShowingInputView):
* Source/WebCore/page/Page.h:
* Source/WebKit/WebProcess/WebPage/WebPage.h:
* Source/WebKit/WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::setIsShowingInputViewForFocusedElement):

Canonical link: https://commits.webkit.org/301057@main