[GFC] Populate eligibility function for layout with some initial logic https://bugs.webkit.org/show_bug.cgi?id=300145 rdar://161924265
Reviewed by Alan Baradlay.

Currently, our canUseForGridLayout eligibility function returns false
whether or not the feature flag is enabled. Now, since we have an
initial implementation that can support some very restricted content, we
should populate this function to allow it. Over time, we will ease out
the conditions here as we claim more and more content to be supported
inside of GFC.

The nature of the content that is currently allowed is essentially fixed
sized grids and items in which the grid is explicitly sized, and the
items are all explicitly placed. There are some additional pieces that
restrict content (e.g. only allows start alignment currently), but this
is the idea at a high level.

Certain AvoidanceReasons are very generic just for this initial set of
conditions but will likely become more fine-grained as we continue to
complete GFC. For example, we return GridHasUnsupportedGridTemplateColumns
in all cases where we run into grids that have any sort of column
styling outside of fixed sizes. Over time, we will likely remove this and
add additional ones that are more descriptive of the condition that
caused us to bail out.

Canonical link: https://commits.webkit.org/301063@main