Wasm catch handler ranges should be inclusive https://bugs.webkit.org/show_bug.cgi?id=299629 rdar://161438523
Reviewed by Yusuke Suzuki.

During generation of catch handlers, we set the ending PC to be the exact position
of the end of the try block, thus creating a [start, end] interval. When we unwind
and check for exception handlers, however, we treat the intervals as [start, end),
meaning that we will not handle edge cases where a CallSiteIndex is exactly at the
end of the interval, thus allowing some exceptions to erroneously propagate.

Test: JSTests/wasm/stress/exception-thrown-at-end-of-catch.js
* JSTests/wasm/stress/exception-thrown-at-end-of-catch.js: Added.
* Source/JavaScriptCore/wasm/WasmHandlerInfo.cpp:
(JSC::Wasm::HandlerInfo::handlerForIndex):

Canonical link: https://commits.webkit.org/301090@main