AX: The Node& variant of AXObjectCache::getOrCreate scopes a renderer CheckedPtr far larger than necessary, which can cause an unnecessary crash https://bugs.webkit.org/show_bug.cgi?id=300254 rdar://161953366
Reviewed by Joshua Hoffman.

Limit the scope of this CheckedPtr to where it's actually used. This avoids an unnecessary crash from side effects later
in the function destroying the renderer (but we don't actually care if it's destroyed, since our use for the renderer
stopped earlier in the function).

Test: accessibility/dynamic-frame-prepend-crash.html

* LayoutTests/accessibility/dynamic-frame-prepend-crash-expected.txt: Added.
* LayoutTests/accessibility/dynamic-frame-prepend-crash.html: Added.
* Source/WebCore/accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::getOrCreateSlow):

Canonical link: https://commits.webkit.org/301101@main