[GTK] Remove support for building against libsoup2 https://bugs.webkit.org/show_bug.cgi?id=300163
Reviewed by Michael Catanzaro.

Remove API 4.0 from the GTK port, which is the only remaining one using
libsoup2. This in turn allows to remove all the USE(SOUP2) guards, and
the SOUP_CHECK_VERSION(2, ...) ones. As side effects, the USE_SOUP2
CMake option is gone, and the definition of ENABLE_SERVER_PRECONNECT
is moved into PlatformEnableGlib.h as it is always enabled for both the
WPE and GTK ports.

Canonical link: https://commits.webkit.org/301131@main