Set serviceClass on WebRTC network connections rdar://149874300 https://bugs.webkit.org/show_bug.cgi?id=291641
Reviewed by Jean-Yves Avenard.

We introduce a feature flag to ocntrol whether marking the service class of a given nw connection (used for WebRTC) as video.
This can help the OS to optimize things for that nw connection.
The flag is off by default for now.
The patch is mostly plumbing the flag from the web process up to nw connections living in networking process.
We do a refactoring to pass the 3 booleans used for creating nw sockets as a structure called RTCSocketCreationFlags.
This makes the code clearer and less error prone.

Canonical link: https://commits.webkit.org/301152@main