Controls clipped in Video Viewer when video is narrower than fullscreen controls https://bugs.webkit.org/show_bug.cgi?id=299895 rdar://147323739
Reviewed by Andy Estes.

While in viewer mode, the video can potentially be narrower than the fullscreen controls.
In this case, the controls are cut off. This patch makes it so when the video is too narrow
to fit the fullscreen controls, we instead show the inline controls. This patch creates a
new MacOSLayoutTraits mode called NarrowViewer. This mode's controls class is MacOSInlineControls.
When we are in this mode, MacOSInlineControls tells the fullscreen button that we are in
fullscreen. Additionally MacOSInlineControls adds a class "narrowviewer" to the controls
element's class list so that in css we only put a background tint on the inline controls
if the controls do not have this class.

* Source/WebCore/Modules/modern-media-controls/controls/layout-traits.js:
(LayoutTraits):
(LayoutTraits.prototype.get mode):
(LayoutTraits.prototype.get isFullscreen):
* Source/WebCore/Modules/modern-media-controls/controls/macos-inline-media-controls.js:
(MacOSInlineMediaControls.):
* Source/WebCore/Modules/modern-media-controls/controls/macos-layout-traits.js:
(MacOSLayoutTraits.prototype.mediaControlsClass):
(MacOSLayoutTraits.prototype.toString):
(MacOSLayoutTraits):
* Source/WebCore/Modules/modern-media-controls/media/media-controller.js:
(MediaController.prototype.get layoutTraits):
(MediaController.prototype._updateControlsIfNeeded):
* Source/WebCore/Modules/modern-media-controls/media/start-support.js:
(StartSupport.prototype._shouldShowStartButton):
(StartSupport):
* Source/WebCore/rendering/cocoa/RenderThemeCocoa.mm:
(WebCore::glassMaterialMediaControlsStyleSheet):

Canonical link: https://commits.webkit.org/301153@main