ASSERTION FAILED: checkedPtrCountWithoutThreadCheck() in SQLiteStorageArea::prepareDatabase https://bugs.webkit.org/show_bug.cgi?id=300294 rdar://162083739
Reviewed by Chris Dumez.

SQLiteStorageArea::prepareDatabase might set m_database to null after m_database is captured in a local CheckedPtr
variable resultDatabase. When the CheckedPtr variable is destroyed at function exit, the assertion will be hit as the
SQLiteDatabase object is already gone. To fix this, this patch uses a local UniqueRef variable to replace the
CheckedPtr. The UniqueRef is only moved to m_database if operation succeeds, so prepareDatabase does not need to reset
m_database in multiple conditions.

This patch also contains a drive-by fix that when prepareDatabase tries to open database for the second time, it does
not pass the right open options.

Canonical link: https://commits.webkit.org/301155@main