[GTK] Fix build for GTK3 after 301078@main https://bugs.webkit.org/show_bug.cgi?id=285379
Reviewed by Adrian Perez de Castro.

After changeset 301078@main, GTK3 build fails with error:

error: 'WEBKIT_WEB_EXTENSION_MATCH_PATTERN_OPTIONS_NONE' was not declared in this scope.

That symbol is defined in 'WebKitWebExtensionMatchPattern.h.in'.

The only file that includes the 'WebKitWebExtensionMatchPattern.h.in' header is
'WebKitWebExtension.h.in'. In 'PlatformGTK.cmake' file 'WebKitWebExtension.h.in'
is only include if 'ENABLE_2022_GLIB_API' is defined.

Thus, 'WebKitWebExtensionMatchPattern.h.in' should only be included in
the same cases 'WebKitWebExtension.h.in' is included.

* Source/WebKit/PlatformGTK.cmake:
* Tools/TestWebKitAPI/glib/CMakeLists.txt:

Canonical link: https://commits.webkit.org/301195@main