Fix Swift demo build on recent toolchains https://bugs.webkit.org/show_bug.cgi?id=300314 rdar://162108584
Reviewed by Richard Robinson.

This makes two changes related to the WTF modulemap so that it builds
OK with recent toolchains. Specifically,
* Two of its headers require objc and that's now noted. (That does
  not fix the build bug here, but seems the right thing to do)
* On Apple internal builds, those headers require headers from
  private frameworks, and that was not previously accommodated by
  the build script.

A note on the build script itself: first, it's only used if
ENABLE_SWIFT_DEMO_URI_SCHEME is turned on, which is not the case
in any EWS or consumer builds. Secondly, this script is expected
to be removed within a few weeks, when the most recent toolchains
have the fixes we need to generate this header as part of the
standard Xcode built-in step. This work is tracked by #50307
although that's presently paused until we get those fixes.

Canonical link: https://commits.webkit.org/301216@main