[Navigation API] imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/entries-after-bfcache-in-iframe.html is constant failure https://bugs.webkit.org/show_bug.cgi?id=300331 rdar://162128068
Reviewed by Basuke Suzuki.

From the test history (linked in https://bugs.webkit.org/show_bug.cgi?id=300331),
we see that this was fixed in 301092@main.

The reason it did not show up in EWS in PR https://github.com/WebKit/WebKit/pull/49361 was because that PR was
uploaded for review before the BFCache was enabled for Navigation API tests
in https://github.com/WebKit/WebKit/pull/51663. So that PR was unable to run this test.

* LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/entries-after-bfcache-in-iframe-expected.txt:
* LayoutTests/platform/ios/TestExpectations:
* LayoutTests/platform/mac/TestExpectations:

Canonical link: https://commits.webkit.org/301228@main