Remove Quirk shouldAvoidPastingImagesAsWebContent for mail.yahoo.* https://bugs.webkit.org/show_bug.cgi?id=279406 rdar://135633486
Reviewed by Brent Fulgham.

Before a drag and drop of images in between an app on iPad to the Yahoo!
Mail compose window, resulted in a blank area. The Yahoo! Mail team
worked on a fix. We have been testing both on iPad and iPhone (with and
without the Quirk) that the feature was working in the current version
of https://mail.yahoo.com/ with success. This quirk is now not required
anymore. Thanks to Yahoo! for their fix.

* Source/WebCore/editing/cocoa/WebContentReaderCocoa.mm:
(WebCore::WebContentReader::readImage):
* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::determineRelevantQuirks):
(WebCore::Quirks::shouldAvoidPastingImagesAsWebContent const): Deleted.
(WebCore::handleYahooQuirks): Deleted.
* Source/WebCore/page/Quirks.h:
* Source/WebCore/page/QuirksData.h:

Canonical link: https://commits.webkit.org/301245@main