Fix integer overflow during FEConvolveMatrix IPC decoder validator https://bugs.webkit.org/show_bug.cgi?id=300303 rdar://161647030
Reviewed by Mike Wyrzykowski.

In the IPC decoder validator for FEConvolveMatrix, the kernelSize()->area()
overflows if given large width and height values. This causes a crash in GPUP.
The fix is to use unclampedArea() which will never overflow.

The fuzzer test case is altered slightly to consume the DidInitialize message
that gets sent back to WebContent to prevent it from reaching the dummy MessageReceiver
and hitting ASSERT_NOT_REACHED()

Test: ipc/decode-feConvolveMatrix-kernelSize-overflow.html
* LayoutTests/ipc/decode-feConvolveMatrix-kernelSize-overflow-expected.txt: Added.
* LayoutTests/ipc/decode-feConvolveMatrix-kernelSize-overflow.html: Added.
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:

Canonical link: https://commits.webkit.org/301246@main